added checks for table creation.
--HG--
branch : v0.0.2

+++ db/migrate/20141020230333_create_user_live_edits.rb
@@ -1,17 +1,21 @@
 class CreateUserLiveEdits < ActiveRecord::Migration
   def self.up
-    create_table :user_live_edits do |t|
-      t.integer :user_id
-      t.string :current_type
-      t.string :current_action
-      t.integer :current_id
-      t.string :current_field
+    if not ActiveRecord::Base.connection.table_exists? 'user_live_editsÏ' then
+      create_table :user_live_edits do |t|
+        t.integer :user_id
+        t.string :current_type
+        t.string :current_action
+        t.integer :current_idÏ
+        t.string :current_field
 
-      t.timestamps
+        t.timestamps
+      end
     end
   end
 
   def self.down
-    drop_table :user_live_edits
+    if  ActiveRecord::Base.connection.table_exists? 'user_live_edits' then
+      drop_table :user_live_edits
+    end
   end
 end
